requirements.txt
mkdocs
pymupdf
pillow